+++ tests/e2e/e2e_bypassminfee_test.go
@@ -5,6 +5,8 @@ import (
 )
 
 func (s *IntegrationTestSuite) testByPassMinFeeWithdrawReward() {
+	// todo gov propose withdraw to be bypass-msg first
+	s.T().Skip()
 	paidFeeAmt := math.LegacyMustNewDecFromStr(minGasPrice).Mul(math.LegacyNewDec(gas)).String()
 	payee := s.chainA.validators[0].keyInfo.GetAddress()
 	// pass