+++ x/globalfee/types/params_test.go
@@ -10,7 +10,8 @@ import (
 func TestDefaultParams(t *testing.T) {
 	p := DefaultParams()
 	require.EqualValues(t, p.MinimumGasPrices, sdk.DecCoins{})
-	require.EqualValues(t, p.BypassMinFeeMsgTypes, DefaultBypassMinFeeMsgTypes())
+	require.EqualValues(t, p.BypassMinFeeMsgTypes, DefaultBypassMinFeeMsgTypes)
+	require.EqualValues(t, p.MaxTotalBypassMinFeeMsgGasUsage, DefaultmaxTotalBypassMinFeeMsgGasUsage)
 }
 
 func Test_validateMinGasPrices(t *testing.T) {